+++ cross_set/cross_set_test.go
@@ -385,7 +385,7 @@ func TestGenAllCrossScores(t *testing.T) {
 	b := board.MakeBoard(board.CrosswordGameBoard)
 	b.SetToGame(alph, board.VsEd)
 
-	GenAllCrossScores(b, nil, dist)
+	GenAllCrossScores(b, dist)
 
 	var testCases = []crossSetTestCase{
 		{8, 8, CrossSetFromString("OS", alph), board.HorizontalDirection, 8},
@@ -412,7 +412,7 @@ func TestGenAllCrossScores(t *testing.T) {
 	}
 	// This one has more nondeterministic (in-between LR) crosssets
 	b.SetToGame(alph, VsMatt)
-	GenAllCrossScores(b, nil, dist)
+	GenAllCrossScores(b, dist)
 	testCases = []crossSetTestCase{
 		{8, 7, CrossSetFromString("S", alph), board.HorizontalDirection, 11},
 		{8, 7, CrossSet(0), board.VerticalDirection, 12},
@@ -441,228 +441,3 @@ func TestGenAllCrossScores(t *testing.T) {
 		}
 	}
 }
-
-/*
-// Copy of TestUpdateCrossSetsForMove with the CrossScoreOnlyGenerator
-func TestUpdateCrossScoresForMove(t *testing.T) {
-	dist, err := alphabet.EnglishLetterDistribution(&DefaultConfig)
-	if err != nil {
-		t.Error(err)
-	}
-	gen := CrossScoreOnlyGenerator{Dist: dist}
-	alph := dist.Alphabet()
-
-	var testCases = []updateCrossesForMoveTestCase{
-		{VsMatt, move.NewScoringMoveSimple(38, "K9", "TAEL", "ABD", alph), "TAEL"},
-		// Test right edge of board
-		{VsMatt2, move.NewScoringMoveSimple(77, "O8", "TENsILE", "", alph), "TENsILE"},
-		// Test through tiles
-		{VsOxy, move.NewScoringMoveSimple(1780, "A1", "OX.P...B..AZ..E", "", alph),
-			"OXYPHENBUTAZONE"},
-		// Test top of board, horizontal
-		{VsJeremy, move.NewScoringMoveSimple(14, "1G", "S.oWED", "D?", alph), "SNoWED"},
-		// Test bottom of board, horizontal
-		{VsJeremy, move.NewScoringMoveSimple(11, "15F", "F..ER", "", alph), "FOYER"},
-	}
-
-	// create a move.
-	for _, tc := range testCases {
-		b := board.MakeBoard(board.CrosswordGameBoard)
-		b.SetToGame(alph, tc.testGame)
-		gen.GenerateAll(b)
-		b.UpdateAllAnchors()
-		b.PlayMove(tc.m, dist)
-		gen.UpdateForMove(b, tc.m)
-		log.Printf(b.ToDisplayText(alph))
-		// Create an identical board, but generate cross-sets for the entire
-		// board after placing the letters "manually".
-		c := board.MakeBoard(board.CrosswordGameBoard)
-		c.SetToGame(alph, tc.testGame)
-		c.PlaceMoveTiles(tc.m)
-		c.TestSetTilesPlayed(c.GetTilesPlayed() + tc.m.TilesPlayed())
-		GenAllCrossScores(c, dist)
-		c.UpdateAllAnchors()
-
-		assert.True(t, b.Equals(c))
-
-		for i, c := range tc.userVisibleWord {
-			row, col, vertical := tc.m.CoordsAndVertical()
-			var rowInc, colInc int
-			if vertical {
-				rowInc = i
-				colInc = 0
-			} else {
-				rowInc = 0
-				colInc = i
-			}
-			uv := b.GetSquare(row+rowInc, col+colInc).Letter().UserVisible(alph)
-			assert.Equal(t, c, uv)
-		}
-	}
-}
-
-// Comparison tests between the two generators for cross-score generation
-
-func compareCrossScores(t *testing.T, b1 *board.GameBoard, b2 *board.GameBoard) {
-	dim := b1.Dim()
-	dirs := []board.BoardDirection{board.HorizontalDirection, board.VerticalDirection}
-	var d board.BoardDirection
-
-	for r := 0; r < dim; r++ {
-		for c := 0; c < dim; c++ {
-			for _, d = range dirs {
-				cs1 := b1.GetCrossScore(r, c, d)
-				cs2 := b2.GetCrossScore(r, c, d)
-				assert.Equal(t, cs1, cs2)
-			}
-		}
-	}
-}
-
-func TestCompareUpdate(t *testing.T) {
-	path := filepath.Join(DefaultConfig.LexiconPath, "gaddag", "America.gaddag")
-	gd, err := gaddag.LoadGaddag(path)
-	if err != nil {
-		t.Error(err)
-	}
-	dist, err := alphabet.EnglishLetterDistribution(&DefaultConfig)
-	if err != nil {
-		t.Error(err)
-	}
-	gen1 := GaddagCrossSetGenerator{Dist: dist, Gaddag: gd}
-	gen2 := CrossScoreOnlyGenerator{Dist: dist}
-	alph := dist.Alphabet()
-
-	var testCases = []updateCrossesForMoveTestCase{
-		{VsMatt, move.NewScoringMoveSimple(38, "K9", "TAEL", "ABD", alph), "TAEL"},
-		// Test right edge of board
-		{VsMatt2, move.NewScoringMoveSimple(77, "O8", "TENsILE", "", alph), "TENsILE"},
-		// Test through tiles
-		{VsOxy, move.NewScoringMoveSimple(1780, "A1", "OX.P...B..AZ..E", "", alph),
-			"OXYPHENBUTAZONE"},
-		// Test top of board, horizontal
-		{VsJeremy, move.NewScoringMoveSimple(14, "1G", "S.oWED", "D?", alph), "SNoWED"},
-		// Test bottom of board, horizontal
-		{VsJeremy, move.NewScoringMoveSimple(11, "15F", "F..ER", "", alph), "FOYER"},
-	}
-
-	// create a move.
-	for _, tc := range testCases {
-		// Run the cross set generator on b1
-		b1 := board.MakeBoard(board.CrosswordGameBoard)
-		b1.SetToGame(alph, tc.testGame)
-		gen1.GenerateAll(b1)
-		b1.UpdateAllAnchors()
-		b1.PlayMove(tc.m, dist)
-		gen1.UpdateForMove(b1, tc.m)
-
-		// Run the cross score generator on b2
-		b2 := board.MakeBoard(board.CrosswordGameBoard)
-		b2.SetToGame(alph, tc.testGame)
-		gen2.GenerateAll(b2)
-		b2.UpdateAllAnchors()
-		b2.PlayMove(tc.m, dist)
-		gen2.UpdateForMove(b2, tc.m)
-
-		compareCrossScores(t, b1, b2)
-	}
-}
-
-func TestCompareGenAll(t *testing.T) {
-	path := filepath.Join(DefaultConfig.LexiconPath, "gaddag", "America.gaddag")
-	gd, err := gaddag.LoadGaddag(path)
-	if err != nil {
-		t.Error(err)
-	}
-	dist, err := alphabet.EnglishLetterDistribution(&DefaultConfig)
-	if err != nil {
-		t.Error(err)
-	}
-	alph := dist.Alphabet()
-
-	var testCases = []board.VsWho{
-		VsEd,
-		VsJeremy,
-		VsMatt,
-		VsMatt2,
-		VsOxy,
-	}
-
-	for _, tc := range testCases {
-		b1 := board.MakeBoard(board.CrosswordGameBoard)
-		b1.SetToGame(alph, tc)
-		GenAllCrossSets(b1, gd, dist)
-
-		b2 := board.MakeBoard(board.CrosswordGameBoard)
-		b2.SetToGame(alph, tc)
-		GenAllCrossScores(b2, dist)
-
-		compareCrossScores(t, b1, b2)
-	}
-}
-
-// Benchmarks
-
-func BenchmarkGenAnchorsAndCrossSets(b *testing.B) {
-	path := filepath.Join(DefaultConfig.LexiconPath, "gaddag", "America.gaddag")
-	gd, err := gaddag.LoadGaddag(path)
-	if err != nil {
-		b.Error(err)
-	}
-	dist, err := alphabet.EnglishLetterDistribution(&DefaultConfig)
-	if err != nil {
-		b.Error(err)
-	}
-	alph := dist.Alphabet()
-
-	bd := board.MakeBoard(board.CrosswordGameBoard)
-	bd.SetToGame(alph, VsOxy)
-	b.ResetTimer()
-
-	// 38 us
-	for i := 0; i < b.N; i++ {
-		bd.UpdateAllAnchors()
-		GenAllCrossSets(bd, gd, dist)
-	}
-}
-
-func BenchmarkMakePlay(b *testing.B) {
-	// Mostly, benchmark the progressive generation of anchors and cross-sets
-	// (as opposed to generating all of them from scratch)
-	path := filepath.Join(DefaultConfig.LexiconPath, "gaddag", "America.gaddag")
-	gd, err := gaddag.LoadGaddag(path)
-	if err != nil {
-		b.Error(err)
-	}
-	dist, err := alphabet.EnglishLetterDistribution(&DefaultConfig)
-	if err != nil {
-		b.Error(err)
-	}
-	gen := GaddagCrossSetGenerator{Dist: dist, Gaddag: gd}
-	alph := dist.Alphabet()
-	bd := board.MakeBoard(board.CrosswordGameBoard)
-	bd.SetToGame(alph, VsMatt)
-	gen.GenerateAll(bd)
-	bd.UpdateAllAnchors()
-
-	// create a move.
-	m := move.NewScoringMove(
-		38,
-		alphabet.MachineWord([]alphabet.MachineLetter{19, 0, 4, 11}), // TAEL
-		alphabet.MachineWord([]alphabet.MachineLetter{0, 1, 3}),
-		true,
-		4,
-		alph,
-		8, 10, "K9")
-
-	b.ResetTimer()
-	// 2.7 us; more than 10x faster than regenerating all anchors every time.
-	// seems worth it.
-	for i := 0; i < b.N; i++ {
-		bd.PlayMove(m, dist)
-		gen.UpdateForMove(bd, m)
-	}
-
-}
-
-*/
